re-organize data splitter/loader tests
